:bug: [FIX]: update db migration files and remove comments

+++ db/migrations/20230531063831_create_users_payment_table.sql
@@ -2,7 +2,7 @@
 CREATE TABLE users_payment (
   id INT PRIMARY KEY AUTO_INCREMENT,
   user_id INT NOT NULL,
-  points INT NOT NULL,
+  points DECIMAL(8,2) NOT NULL,
   INDEX user_id (user_id),
   FOREIGN KEY (user_id) REFERENCES users (id)
 );
